refactor: from demoapp containerize the events listener
