+++ webapps/svelte/src/lib/app/LoginApp.ts
@@ -3,6 +3,7 @@ import { AuthenticationProtocol } from "../api/AuthenticationGW";
 import { log } from "../utils/helpers";
 import AuthService from "./AuthService";
 import AuthState from "./AuthState";
+import LoginEventListener from "./LoginEventListener";
 import UIState from "./UIState";
 
 export default class LoginApp {
@@ -44,49 +45,6 @@ export default class LoginApp {
 
   async engage() {
     this.initializeComponents();
-    await this.engageLoginFlowEventsListener();
-  }
-
-  /**************************************************************************
-  * 
-  * Event listeners
-  * 
-  /**************************************************************************/
-  async engageLoginFlowEventsListener() {
-    const urlParams = new URLSearchParams(window.location.search);
-
-    const affectsThisApp = urlParams.has("provider") && urlParams.get("provider").toLowerCase() === this.getName().toLowerCase();
-    if (!this.AuthState.isLoggedIn()) {
-      if (affectsThisApp && urlParams.has("loginCode")) {
-        this.log("LoginFlowEventsListener", "Login code received, fetching auth token..");
-        //
-        // Handle login response
-        //
-        const loginCode = urlParams.get("loginCode");
-        try {
-          const tokens = await this.AuthService.fetchAuthTokens(loginCode);
-          this.AuthState.login(tokens); // Store token in local storage
-          await this.AuthState.handleLoggedIn(); // Fetch user info
-          this.UIState.resetViewState(); // reset view state
-        } catch (error) {
-          this.log("LoginFlowEventsListener", "Failed to fetch auth token", error);
-        }
-      } else {
-        this.UIState.handleCurrentState(); // Init UI
-      }
-    } else if (affectsThisApp && urlParams.has("logout")) {
-      this.log("LoginFlowEventsListener", "Logout event received, logging out");
-      //
-      // Handle logout response
-      //
-      const logoutResponse = urlParams.get("logout");
-      if (logoutResponse === "success") {
-        this.AuthState.logout();
-        this.UIState.resetViewState(); // reset view state
-      }
-    } else {
-      await this.AuthState.handleLoggedIn(); // Validate login
-      this.UIState.handleCurrentState(); // Update UI
-    }
+    await LoginEventListener(this); // Listen for auth events
   }
 }
